Converted the BasicNumeric interface to a class
