--- core/src/main/java/dk/alexandra/fresco/suite/dummy/arithmetic/DummyArithmetic.java
@@ -1,54 +1,0 @@
-/*
- * Copyright (c) 2017 FRESCO (http://github.com/aicis/fresco).
- *
- * This file is part of the FRESCO project.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
- * associated documentation files (the "Software"), to deal in the Software without restriction,
- * including without limitation the rights to use, copy, modify, merge, publish, distribute,
- * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or
- * substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
- * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
- * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- *
- * FRESCO uses SCAPI - http://crypto.biu.ac.il/SCAPI, Crypto++, Miracl, NTL, and Bouncy Castle.
- * Please see these projects for any further licensing issues.
- *******************************************************************************/
-
-package dk.alexandra.fresco.suite.dummy.arithmetic;
-
-import dk.alexandra.fresco.lib.field.integer.BasicNumeric;
-import java.math.BigInteger;
-
-/**
- * Implementation of {@link BasicNumeric} for the Dummy Arithmetic suite.
- */
-public class DummyArithmetic implements BasicNumeric {
-
-  private BigInteger mod;
-  private int maxBitLength;
-
-  public DummyArithmetic(BigInteger mod, int maxBitLength) {
-    this.mod = mod;
-    this.maxBitLength = maxBitLength;
-  }
-
-
-  @Override
-  public int getMaxBitLength() {
-    return maxBitLength;
-  }
-
-  @Override
-  public BigInteger getModulus() {
-    return mod;
-  }
-
-}
