+++ core/src/test/java/dk/alexandra/fresco/lib/math/polynomial/TestPolynomial.java
@@ -4,7 +4,7 @@ import dk.alexandra.fresco.framework.Computation;
 import dk.alexandra.fresco.framework.builder.numeric.BuilderFactoryNumeric;
 import dk.alexandra.fresco.framework.builder.numeric.NumericBuilder;
 import dk.alexandra.fresco.framework.value.SInt;
-import dk.alexandra.fresco.suite.dummy.arithmetic.DummyArithmetic;
+import dk.alexandra.fresco.lib.field.integer.BasicNumeric;
 import dk.alexandra.fresco.suite.dummy.arithmetic.DummyArithmeticBuilderFactory;
 import java.math.BigInteger;
 import java.util.Arrays;
@@ -18,31 +18,31 @@ public class TestPolynomial {
 
   @Test(expected = IndexOutOfBoundsException.class)
   public void testPolynomial() {
-    DummyArithmetic dummyFact = new DummyArithmetic(BigInteger.valueOf(1001), 8);
-    BuilderFactoryNumeric builderFactory =new DummyArithmeticBuilderFactory(dummyFact);
+    BasicNumeric dummyFact = new BasicNumeric(8, BigInteger.valueOf(1001));
+    BuilderFactoryNumeric builderFactory = new DummyArithmeticBuilderFactory(dummyFact);
     NumericBuilder numeric = builderFactory.createNumericBuilder(builderFactory.createSequential());
 
-    int[] coefficients = new int[] {1,2,3,4};
-    
+    int[] coefficients = new int[]{1, 2, 3, 4};
+
     List<Computation<SInt>> secretCoefficients =
         Arrays.stream(coefficients)
             .mapToObj(BigInteger::valueOf)
             .map(numeric::known)
             .collect(Collectors.toList());
-    
+
     Polynomial p = new PolynomialImpl(secretCoefficients);
-    
+
     Assert.assertThat(p.getMaxDegree(), Is.is(4));
-   // for(int i = 0; i< 4; i++) {
+    // for(int i = 0; i< 4; i++) {
     //  p.setCoefficient(i, new DummyArithmeticSInt(i));  
-   // }
-    
-    for(int i = 0; i< 4; i++) {
-      Assert.assertThat(p.getCoefficient(i), Is.is(secretCoefficients.get(i)));  
-    }    
-    
-      p.getCoefficient(5);
+    // }
+
+    for (int i = 0; i < 4; i++) {
+      Assert.assertThat(p.getCoefficient(i), Is.is(secretCoefficients.get(i)));
+    }
+
+    p.getCoefficient(5);
   }
-  
+
 }
