--- suite/spdz/src/main/java/dk/alexandra/fresco/suite/spdz/utils/SpdzNumeric.java
@@ -1,59 +1,0 @@
-/*
- * Copyright (c) 2015, 2016 FRESCO (http://github.com/aicis/fresco).
- *
- * This file is part of the FRESCO project.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.  IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- *
- * FRESCO uses SCAPI - http://crypto.biu.ac.il/SCAPI, Crypto++, Miracl, NTL,
- * and Bouncy Castle. Please see these projects for any further licensing issues.
- *******************************************************************************/
-package dk.alexandra.fresco.suite.spdz.utils;
-
-import dk.alexandra.fresco.lib.field.integer.BasicNumeric;
-import java.math.BigInteger;
-
-public class SpdzNumeric implements BasicNumeric {
-
-  private final int maxBitLength;
-  private final BigInteger modulus;
-
-  /**
-   * @param maxBitLength The maximum length in bits that the numbers in the application will have.
-   * If you have greater knowledge of your application, you can create several factorys, each with a
-   * different maxBitLength to increase performance.
-   * @param modulus the modules used in the application
-   */
-  public SpdzNumeric(int maxBitLength, BigInteger modulus) {
-    this.maxBitLength = maxBitLength;
-    this.modulus = modulus;
-  }
-
-  @Override
-  public int getMaxBitLength() {
-    return this.maxBitLength;
-  }
-
-
-  @Override
-  public BigInteger getModulus() {
-    return modulus;
-  }
-
-}
